Update github-stats status checks (#455)
# Pull Request

## Description

This pull request updates the `github-stats.tf` file to enhance the
default branch protection checks by adding new security and
vulnerability scanning tools.

### Enhancements to default branch protection:

*
[`stack/github-stats.tf`](diffhunk://#diff-4c83331edafb5463d9be52eca7c9e213c0868cddd414c2b7d97d4e1559a70ab1R66-R68):
Added three new checks to the list of common code checks: "Check for
Secrets with Gitleaks," "Check for Secrets with TruffleHog," and "Check
for Vulnerabilities with Grype." These additions improve the security
and reliability of the codebase by automating the detection of secrets
and vulnerabilities.