Rename dev-cli to dx-cli (#498)
# Pull Request

## Description

This pull request renames the Terraform configuration for a GitHub
repository from `dev-cli` to `dx-cli` and updates all references
accordingly. The changes ensure consistency in naming and maintain
functionality for the repository's configuration.

### File renaming and resource updates:

* Renamed the file `stack/dev-cli.tf` to `stack/dx-cli.tf` and updated
the resource name from `github_repository.dev-cli` to
`github_repository.dx-cli`. This includes changing the `name` attribute
of the repository from `dev-cli` to `dx-cli`.

### Updates to associated resources and modules:

* Updated the `github_repository_dependabot_security_updates` resource
to reflect the new repository name (`dx-cli`) in both its declaration
and the `repository` attribute.
* Updated the `module "default_branch_protection"` configuration to use
the new repository name (`dx-cli`) in the `repository_name` and
`depends_on` attributes.