Update remove-pr-bot-collaborators description (#523)
# Pull Request

## Description


This pull request makes small but important updates to repository
configuration and metadata. The main changes are a clarification of the
repository description for `remove-pr-bot-collaborators` and a minor
adjustment to required code scanning tools and status checks for the
default branch protection in `github-stats`.

Repository metadata improvement:

* Updated the `description` field for the `remove-pr-bot-collaborators`
repository to clearly state its purpose as a bot for automatically
removing collaborators from pull requests.

Default branch protection updates:

* Removed the `"Test TypeScript Code"` status check from the required
status checks list for the `github-stats` repository.
* Removed `"SonarCloud"` from the list of required code scanning tools
for the `github-stats` repository.