Update github-stats-analyser status checks (#559)
# Pull Request

## Description

This pull request makes a minor update to the default branch protection
rules for the `github-stats-analyser` module. The most notable change is
the removal of the "CodeQL Analysis (go) / Analyse code" check from the
required status checks.

* Removed the "CodeQL Analysis (go) / Analyse code" status check from
the list of required checks in the
`github-stats-analyser_default_branch_protection` module
(`stack/github-stats-analyser.tf`).