Update repository-visualiser status checks (#582)
# Pull Request

## Description

This pull request updates the configuration for default branch
protection in the `repository-visualiser` stack. The main change is to
centralize and simplify the management of required status checks by
concatenating a shared list of checks with repository-specific ones.

Branch protection configuration improvements:

* Changed the `required_status_checks` assignment in
`stack/repository-visualiser.tf` to use the `concat` function, combining
a repository-specific list with the shared
`local.common_required_status_checks` variable. This makes it easier to
maintain and update common checks across multiple repositories.