Update Terraform Lockfile (#612)
# Pull Request

## Description

This pull request makes a minor update to the Terraform lock file,
specifically for the GitHub provider. The change adds a new hash entry
to ensure the integrity and verification of the provider version.

* Added a new hash (`h1:GV6m1zdKNV0ECNhvzj7UoBXQz1Cv2UQa0UZ7Wt4RgTw=`)
to the `provider "registry.opentofu.org/integrations/github"` block in
`stack/.terraform.lock.hcl` for version verification.