Update github-stats status checks (#634)
# Pull Request

## Description

This pull request makes a minor update to the required status checks for
the GitHub stats module. It adds the "Run TypeScript Format Checks" job
to the list of required checks, ensuring that TypeScript code formatting
is enforced before merging.

* Added "Run TypeScript Format Checks" to the required status checks in
the `github-stats_default_branch_protection` module in
`stack/github-stats.tf`.