Update Tofu Provider Version (#636)
# Pull Request

## Description

This pull request updates the version of the `github` Terraform provider
from `6.6.0` to `6.7.0` across the codebase to ensure compatibility with
the latest features and bug fixes. It also updates the provider lock
file to reflect the new version and its associated hashes.

Provider version upgrade:

* Updated the `github` provider version from `6.6.0` to `6.7.0` in
`modules/default-branch-protection/version.tf` and `stack/terraform.tf`
to use the latest provider release.
[[1]](diffhunk://#diff-083dcdd2848fb453f0966cc360d8160b73b9759560863e07dc0e7ae21f851b8aL7-R7)
[[2]](diffhunk://#diff-071e1b4d3f4377396546dc5bc41672a414c0723beedc4b4ed2a37a2c5086ccf5L6-R6)

Dependency lock update:

* Updated `stack/.terraform.lock.hcl` to lock the `github` provider at
version `6.7.0` and replaced the provider hashes to match the new
version.