Update Checkov Version (#671)
# Pull Request

## Description

This pull request makes a minor update to the code quality workflow by
upgrading the Checkov GitHub Action to a newer version. This ensures the
latest security and compliance checks are used in CI.

- DevOps/CI improvements:
* Upgraded the `bridgecrewio/checkov-action` in
`.github/workflows/code-checks.yml` from version `v12.3067.0` to
`v12.3075.0` to use the latest Checkov features and fixes.